Move task.py out of backend/core.
It's a core pants concept and not part of what we
usually mean by "backend".  This is a step one of
sorting out the mess in backend/core.

Points all pants code to the new location of task.py,
which is a new top-level 'task' package. A future change
will move ConsoleTask and other task-related stuff to there.

There are forwarding aliases in the original task.py, with a
deprecation warning, so external plugins shouldn't break.

Testing Done:
CI passes: https://travis-ci.org/pantsbuild/pants/builds/91325801

Reviewed at https://rbcommons.com/s/twitter/r/3130/
