+++ contrib/cpp/src/python/pants/contrib/cpp/tasks/cpp_task.py
@@ -7,8 +7,8 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 
 import subprocess
 
-from pants.backend.core.tasks.task import Task
 from pants.base.exceptions import TaskError
+from pants.task.task import Task
 
 from pants.contrib.cpp.targets.cpp_binary import CppBinary
 from pants.contrib.cpp.targets.cpp_library import CppLibrary
