+++ src/python/pants/backend/codegen/tasks/BUILD
@@ -18,8 +18,8 @@ python_library(
   name = 'common',
   sources = ['__init__.py'],
   dependencies = [
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/base:exceptions',
+    'src/python/pants/task',
   ]
 )
 
@@ -115,11 +115,11 @@ python_library(
   sources = ['simple_codegen_task.py'],
   dependencies = [
     '3rdparty/python/twitter/commons:twitter.common.collections',
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:exceptions',
     'src/python/pants/base:workunit',
     'src/python/pants/build_graph',
+    'src/python/pants/task',
     'src/python/pants/util:dirutil',
   ],
 )
