+++ src/python/pants/backend/core/tasks/BUILD
@@ -1,6 +1,25 @@
 # Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
+# =======>
+# TODO: Remove these after task.py's deprecation cycle is over.
+target(
+  name = 'common',
+  dependencies = [
+    'src/python/pants/task',
+  ]
+)
+
+python_library(
+  name = 'task',
+  sources = ['task.py'],
+  dependencies = [
+    'src/python/pants/base:deprecated',
+    'src/python/pants/task',
+  ],
+)
+# <======
+
 target(
   name = 'all',
   dependencies = [
@@ -9,7 +28,6 @@ target(
     ':changed_target_goals',
     ':clean',
     ':cloc',
-    ':common',
     ':confluence_publish',
     ':console_task',
     ':deferred_sources_mapper',
@@ -34,17 +52,8 @@ target(
     ':scm_publish',
     ':sorttargets',
     ':targets_help',
-    ':task',
     ':test_task_mixin',
     ':what_changed',
-  ]
-)
-
-target(
-  name = 'common',
-  dependencies = [
-    ':task',
-    'src/python/pants/base:exceptions',
   ]
 )
 
@@ -56,12 +65,12 @@ python_library(
   ],
   dependencies = [
     ':console_task',
-    ':task',
     'src/python/pants/base:exceptions',
     'src/python/pants/base:generator',
     'src/python/pants/goal',
     'src/python/pants/help',
     'src/python/pants/option',
+    'src/python/pants/task',
   ],
 )
 
@@ -77,9 +86,7 @@ python_library(
     ':builddictionary_resources',
   ],
   dependencies = [
-    ':common',
     ':reflect',
-    ':task',
     '3rdparty/python:docutils',
     '3rdparty/python:setuptools',
     '3rdparty/python/twitter/commons:twitter.common.collections',
@@ -91,6 +98,7 @@ python_library(
     'src/python/pants/base:generator',
     'src/python/pants/build_graph',
     'src/python/pants/goal',
+    'src/python/pants/task',
     'src/python/pants/util:dirutil',
   ],
 )
@@ -105,8 +113,8 @@ python_library(
   name = 'clean',
   sources = ['clean.py'],
   dependencies = [
-    ':task',
     'src/python/pants/base:build_environment',
+    'src/python/pants/task',
     'src/python/pants/util:dirutil',
   ],
 )
@@ -129,10 +137,10 @@ python_library(
   sources = ['confluence_publish.py'],
   dependencies = [
     '3rdparty/python/twitter/commons:twitter.common.confluence',
-    ':task',
     'src/python/pants/backend/core/targets:common',
     'src/python/pants/base:exceptions',
     'src/python/pants/binaries:binary_util',
+    'src/python/pants/task',
     'src/python/pants/util:dirutil',
   ],
 )
@@ -141,7 +149,7 @@ python_library(
   name = 'console_task',
   sources = ['console_task.py'],
   dependencies = [
-    ':common',
+    'src/python/pants/task',
     'src/python/pants/util:dirutil',
   ],
 )
@@ -150,9 +158,9 @@ python_library(
   name = 'deferred_sources_mapper',
   sources= ['deferred_sources_mapper.py'],
   dependencies = [
-    ':task',
     'src/python/pants/base:payload_field',
     'src/python/pants/build_graph',
+    'src/python/pants/task',
   ],
 )
 
@@ -203,11 +211,11 @@ python_library(
   sources = ['group_task.py'],
   dependencies = [
     '3rdparty/python:six',
-    ':task',
     'src/python/pants/base:workunit',
     'src/python/pants/build_graph',
     'src/python/pants/goal',
     'src/python/pants/option',
+    'src/python/pants/task',
   ],
 )
 
@@ -240,7 +248,6 @@ python_library(
     '3rdparty/python:Pygments',
     '3rdparty/python:docutils',
     '3rdparty/python:six',
-    ':task',
     'src/python/pants/backend/core/targets:common',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:exceptions',
@@ -248,6 +255,7 @@ python_library(
     'src/python/pants/base:workunit',
     'src/python/pants/binaries:binary_util',
     'src/python/pants/build_graph',
+    'src/python/pants/task',
     'src/python/pants/util:dirutil',
   ],
 )
@@ -269,8 +277,8 @@ python_library(
   name = 'mutex_task_mixin',
   sources = ['mutex_task_mixin.py'],
   dependencies = [
-    ':task',
     'src/python/pants/base:exceptions',
+    'src/python/pants/task',
   ]
 )
 
@@ -278,7 +286,7 @@ python_library(
   name = 'noop',
   sources = ['noop.py'],
   dependencies = [
-    ':common',
+    'src/python/pants/task',
   ],
 )
 
@@ -314,11 +322,11 @@ python_library(
   name = 'reporting_server',
   sources = ['reporting_server.py'],
   dependencies = [
-    ':task',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:run_info',
     'src/python/pants/binaries:binary_util',
     'src/python/pants/reporting',
+    'src/python/pants/task',
   ],
 )
 
@@ -350,9 +358,9 @@ python_library(
   name = 'run_prep_command',
   sources = ['run_prep_command.py'],
   dependencies = [
-    ':task',
     'src/python/pants/base:exceptions',
     'src/python/pants/base:workunit',
+    'src/python/pants/task',
   ],
 )
 
@@ -370,9 +378,9 @@ python_library(
   name = 'sorttargets',
   sources = ['sorttargets.py'],
   dependencies = [
-    ':common',
     ':console_task',
     'src/python/pants/build_graph',
+    'src/python/pants/task',
   ],
 )
 
@@ -380,8 +388,8 @@ python_library(
   name = 'target_filter_task_mixin',
   sources = ['target_filter_task_mixin.py'],
   dependencies = [
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/base:exceptions',
+    'src/python/pants/task',
   ]
 )
 
@@ -392,8 +400,6 @@ python_library(
     ':targets_help_resources',
   ],
   dependencies = [
-    ':builddictionary',
-    ':common',
     ':console_task',
     ':reflect',
   ],
@@ -405,29 +411,11 @@ resources(
 )
 
 python_library(
-  name = 'task',
-  sources = ['task.py'],
-  dependencies = [
-    '3rdparty/python/twitter/commons:twitter.common.collections',
-    'src/python/pants/base:exceptions',
-    'src/python/pants/base:hash_utils',
-    'src/python/pants/base:worker_pool',
-    'src/python/pants/base:workunit',
-    'src/python/pants/cache',
-    'src/python/pants/invalidation',
-    'src/python/pants/ivy',
-    'src/python/pants/java:executor',
-    'src/python/pants/reporting',
-    'src/python/pants/subsystem',
-    'src/python/pants/util:meta',
-  ],
-)
-
-python_library(
   name = 'test_task_mixin',
   sources = ['test_task_mixin.py'],
   dependencies = [
-    ':task',
+    'src/python/pants/base:exceptions',
+    'src/python/pants/util:timeout',
   ],
 )
