+++ src/python/pants/backend/core/tasks/deferred_sources_mapper.py
@@ -7,9 +7,9 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 
 import logging
 
-from pants.backend.core.tasks.task import Task
 from pants.base.payload_field import DeferredSourcesField
 from pants.build_graph.address_lookup_error import AddressLookupError
+from pants.task.task import Task
 
 
 logger = logging.getLogger(__name__)
