+++ src/python/pants/backend/core/tasks/mutex_task_mixin.py
@@ -8,8 +8,8 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 from abc import abstractmethod
 from collections import defaultdict
 
-from pants.backend.core.tasks.task import Task
 from pants.base.exceptions import TaskError
+from pants.task.task import Task
 
 
 class MutexTaskMixin(Task):
