+++ src/python/pants/backend/jvm/tasks/BUILD
@@ -46,10 +46,10 @@ python_library(
   dependencies = [
     ':jvm_task',
     ':jvm_tool_task_mixin',
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/base:workunit',
     'src/python/pants/java:util',
+    'src/python/pants/task',
   ],
 )
 
@@ -190,7 +190,6 @@ python_library(
   name = 'ivy_task_mixin',
   sources = ['ivy_task_mixin.py'],
   dependencies = [
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/backend/jvm:ivy_utils',
     'src/python/pants/backend/jvm:jar_dependency_utils',
@@ -200,6 +199,7 @@ python_library(
     'src/python/pants/invalidation',
     'src/python/pants/ivy',
     'src/python/pants/java:util',
+    'src/python/pants/task',
     'src/python/pants/util:memo',
   ],
 )
@@ -306,7 +306,6 @@ python_library(
     ':classpath_util',
     ':jvm_task',
     ':jvm_tool_task_mixin',
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/backend/jvm/subsystems:shader',
     'src/python/pants/backend/jvm/targets:java',
     'src/python/pants/backend/jvm/targets:jvm',
@@ -315,6 +314,7 @@ python_library(
     'src/python/pants/base:workunit',
     'src/python/pants/binaries:binary_util',
     'src/python/pants/java:util',
+    'src/python/pants/task',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:dirutil',
     'src/python/pants/util:xml_parser',
@@ -344,12 +344,12 @@ python_library(
   dependencies = [
     '3rdparty/python/twitter/commons:twitter.common.collections',
     ':classpath_util',
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/backend/jvm/targets:scala',
     'src/python/pants/base:build_environment',
     'src/python/pants/build_graph',
     'src/python/pants/java/distribution',
+    'src/python/pants/task',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:memo',
   ]
@@ -361,7 +361,6 @@ python_library(
   dependencies = [
     ':jvm_dependency_analyzer',
     '3rdparty/python/twitter/commons:twitter.common.collections',
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:exceptions',
     'src/python/pants/backend/jvm/tasks:ivy_task_mixin',
@@ -369,6 +368,7 @@ python_library(
     'src/python/pants/backend/jvm/targets:scala',
     'src/python/pants/build_graph',
     'src/python/pants/java/distribution',
+    'src/python/pants/task',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:memo',
   ],
@@ -392,9 +392,9 @@ python_library(
   dependencies = [
     '3rdparty/python:ansicolors',
     'src/python/pants/backend/core/tasks:console_task',
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/base:exceptions',
     'src/python/pants/backend/jvm/targets:jvm',
+    'src/python/pants/task',
     'src/python/pants/util:memo',
   ]
 )
@@ -404,13 +404,13 @@ python_library(
   sources = ['jvm_run.py'],
   dependencies = [
     ':jvm_task',
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/base:workunit',
     'src/python/pants/fs',
     'src/python/pants/java:executor',
     'src/python/pants/java/distribution',
     'src/python/pants/java:util',
+    'src/python/pants/task',
     'src/python/pants/util:dirutil',
     'src/python/pants/util:strutil',
   ],
@@ -421,8 +421,8 @@ python_library(
   sources = ['jvm_task.py'],
   dependencies = [
     ':classpath_util',
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/backend/jvm/subsystems:jvm',
+    'src/python/pants/task',
   ],
 )
 
@@ -430,9 +430,9 @@ python_library(
   name = 'jvm_tool_task_mixin',
   sources = ['jvm_tool_task_mixin.py'],
   dependencies = [
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/backend/jvm/subsystems:jvm_tool_mixin',
     'src/python/pants/base:exceptions',
+    'src/python/pants/task',
   ],
 )
 
@@ -443,7 +443,7 @@ python_library(
     ':jvm_task',
     'src/python/pants/base:exceptions',
     'src/python/pants/binaries:binary_util',
-    'src/python/pants/util:dirutil',\
+    'src/python/pants/util:dirutil',
   ],
 )
 
@@ -452,13 +452,13 @@ python_library(
   sources = ['nailgun_task.py'],
   dependencies = [
     ':jvm_tool_task_mixin',
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/base:exceptions',
     'src/python/pants/java/distribution:distribution',
     'src/python/pants/java:executor',
     'src/python/pants/java:nailgun_executor',
     'src/python/pants/java:util',
+    'src/python/pants/task',
   ],
 )
 
@@ -498,8 +498,8 @@ python_library(
   name = 'resources_task',
   sources = ['resources_task.py'],
   dependencies = [
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/option',
+    'src/python/pants/task',
   ],
 )
 
@@ -545,11 +545,11 @@ python_library(
   sources = ['unpack_jars.py'],
   dependencies = [
     '3rdparty/python/twitter/commons:twitter.common.dirutil',
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/backend/jvm/tasks:jar_import_products',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:fingerprint_strategy',
     'src/python/pants/fs',
+    'src/python/pants/task',
   ]
 )
