+++ src/python/pants/backend/jvm/tasks/jvm_compile/BUILD
@@ -49,7 +49,7 @@ python_library(
   sources = ['jvm_classpath_publisher.py'],
   dependencies = [
     ':jvm_compile',
-    'src/python/pants/backend/core/tasks:task',
+    'src/python/pants/task',
     'src/python/pants/util:dirutil',
     'src/python/pants/util:fileutil',
   ]
