+++ src/python/pants/backend/jvm/tasks/nailgun_task.py
@@ -7,7 +7,6 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 
 import os
 
-from pants.backend.core.tasks.task import Task, TaskBase
 from pants.backend.jvm.targets.jar_dependency import JarDependency
 from pants.backend.jvm.tasks.jvm_tool_task_mixin import JvmToolTaskMixin
 from pants.base.exceptions import TaskError
@@ -15,6 +14,7 @@ from pants.java import util
 from pants.java.distribution.distribution import DistributionLocator
 from pants.java.executor import SubprocessExecutor
 from pants.java.nailgun_executor import NailgunExecutor, NailgunProcessGroup
+from pants.task.task import Task, TaskBase
 
 
 class NailgunTaskBase(JvmToolTaskMixin, TaskBase):
