+++ src/python/pants/backend/python/tasks/BUILD
@@ -16,7 +16,6 @@ python_library(
     'src/python/pants/backend/codegen/targets:python',
     'src/python/pants/backend/core/targets:common',
     'src/python/pants/backend/core/tasks:repl_task_mixin',
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/backend/core/tasks:test_task_mixin',
     'src/python/pants/backend/python/targets:python',
     'src/python/pants/backend/python:antlr_builder',
@@ -34,6 +33,7 @@ python_library(
     'src/python/pants/binaries:thrift_util',
     'src/python/pants/ivy',
     'src/python/pants/option',
+    'src/python/pants/task',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:dirutil',
     'src/python/pants/util:memo',
