+++ src/python/pants/bin/BUILD
@@ -10,7 +10,6 @@ python_library(
     '3rdparty/python:pex',
     '3rdparty/python:setuptools',
     '3rdparty/python/twitter/commons:twitter.common.collections',
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/backend/jvm/tasks:nailgun_task',
     'src/python/pants/backend/python:python_setup',
     'src/python/pants/base:build_environment',
@@ -29,6 +28,7 @@ python_library(
     'src/python/pants/option',
     'src/python/pants/reporting',
     'src/python/pants/subsystem',
+    'src/python/pants/task',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:dirutil',
     'src/python/pants/util:filtering',
