+++ tests/python/pants_test/backend/core/tasks/BUILD
@@ -154,8 +154,8 @@ python_tests(
   name='task',
   sources=['test_task.py'],
   dependencies=[
-    'src/python/pants/backend/core/tasks:task',
     'src/python/pants/build_graph',
+    'src/python/pants/task',
     'tests/python/pants_test/tasks:task_test_base',
   ]
 )
