+++ tests/python/pants_test/backend/core/tasks/test_test_task_mixin.py
@@ -7,9 +7,9 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 
 from mock import patch
 
-from pants.backend.core.tasks.task import TaskBase
 from pants.backend.core.tasks.test_task_mixin import TestTaskMixin
 from pants.base.exceptions import TestFailedTaskError
+from pants.task.task import TaskBase
 from pants.util.timeout import TimeoutReached
 from pants_test.tasks.task_test_base import TaskTestBase
