+++ tests/python/pants_test/backend/jvm/tasks/test_bootstrap_jvm_tools.py
@@ -9,7 +9,6 @@ import os
 import subprocess
 from contextlib import contextmanager
 
-from pants.backend.core.tasks.task import Task
 from pants.backend.jvm.subsystems.shader import Shading
 from pants.backend.jvm.targets.jar_dependency import JarDependency
 from pants.backend.jvm.targets.jar_library import JarLibrary
@@ -17,6 +16,7 @@ from pants.backend.jvm.tasks.bootstrap_jvm_tools import BootstrapJvmTools
 from pants.backend.jvm.tasks.jvm_tool_task_mixin import JvmToolTaskMixin
 from pants.java.distribution.distribution import DistributionLocator
 from pants.java.executor import SubprocessExecutor
+from pants.task.task import Task
 from pants.util.contextutil import open_zip
 from pants_test.jvm.jvm_tool_task_test_base import JvmToolTaskTestBase
 from pants_test.subsystem.subsystem_util import subsystem_instance
